requirements.txt
# Physics-Informed Neural Networks for Wind Flow Simulation
# Python 3.9+ required

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Scientific Computing with Dask for large-scale operations
dask[array]>=2023.5.0

# Machine Learning & Data Processing
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# 3D Geometry Handling
numpy-stl>=3.0.0
vtk>=9.2.0

# System Utilities
psutil>=5.9.0

# GPU Monitoring (NVIDIA only - optional)
GPUtil>=1.4.0
nvidia-ml-py3>=7.352.0  # pynvml

# Development Tools (Optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Note: Data files should be downloaded separately from:
# https://gitlab.cern.ch/abinakbe/pinns/-/tree/master/data
